Settings resizing from base
edited resizing info so that it all works from a base position and size and scales reletive to that. Was able to fix all resizing issues with this change however makes menus a lot more memory intensive. Pointer positions are now pointer values so they dont need to be reasined when window size changes